# Conflicts:
#	app/src/main/java/com/jetsynthesys/rightlife/ai_package/data/repository/ApiService.kt

+++ app/src/main/java/com/jetsynthesys/rightlife/ai_package/model/UnsyncedWorkout.kt
@@ -31,7 +31,7 @@ data class UnsyncedWorkout(
     val activityFactor: Float,
 
     @SerializedName("calories_burned")
-    val caloriesBurned: Int,
+    val caloriesBurned: Double,
 
     @SerializedName("timestamp")
     val timestamp: String,
